admin_panel/requirements/dev.txt
-r common.txt

django-debug-toolbar==1.6